Cleanup, add Delete board validation

+++ src/backend/Services/Board/Board.Api/Features/Board/DeleteBoard/DeleteBoardEndpoint.cs
@@ -1,4 +1,5 @@
 using Board.Application.Abstractions.Repositories;
+using Board.Application.Abstractions.Services;
 using FastEndpoints;
 
 namespace Board.Api.Features.Board.DeleteBoard;
@@ -11,9 +12,11 @@ public class DeleteBoardEndpoint : EndpointWithoutRequest
     }
 
     private readonly IBoardRepository _repository;
-    public DeleteBoardEndpoint(IBoardRepository repository)
+    private readonly ICurrentUserProvider _currentUserProvider;
+    public DeleteBoardEndpoint(IBoardRepository repository, ICurrentUserProvider currentUserProvider)
     {
         _repository = repository;
+        _currentUserProvider = currentUserProvider;
     }
 
     public override async Task HandleAsync(CancellationToken cancellationToken)
@@ -21,6 +24,12 @@ public class DeleteBoardEndpoint : EndpointWithoutRequest
         Guid id = Route<Guid>("id");
 
         Domain.Entities.Board entity = await _repository.GetAsync(x => x.Id == id, cancellationToken);
+
+        Domain.Entities.BoardUser boardOwner = await _repository.GetBoardOwnerAsync(entity.Id, cancellationToken);
+        if (boardOwner.Email != _currentUserProvider.GetCurrentUserEmail())
+        {
+            throw new Exception("You can not remove this board");
+        }
 
         await _repository.DeleteAsync(entity, cancellationToken);
